chore: TDET-50 Renamed files as per team conventions
