+++ dags/brz_bonds_daily/uploaders.py
@@ -2,11 +2,11 @@ import json
 
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
 
-from brz_bonds_meta_monthly.brz_bonds_meta_constants import S3_BUCKET
+from brz_bonds_daily.constants import S3_BUCKET
 
 
-# Bonds meta data uploader
-def upload_bonds_metadata_to_s3(payload, key):
+# Bonds uploader: this one is not a task.
+def upload_bonds_to_s3(payload, key):
     s3 = S3Hook(aws_conn_id="aws_conn_id")
     s3.load_string(
         string_data=json.dumps(payload),
